Create a new demo: Wave_CDClipmap_Test, but it not perfect.

+++ postprocessingcore/src/main/java/jet/opengl/postprocessing/util/Rectf.java
@@ -33,7 +33,7 @@ public class Rectf {
         height =o.height;
     }
 
-    public void set(int x, int y, int width, int height){
+    public void set(float x, float y, float width, float height){
         this.x =x;
         this.y =y;
         this.width =width;
